Update Double_sentinel_list.h
in erase(), you have to decrement list_size accordingly. Otherwise, size 0 and empty 1 in assign result in errors. I changed it.
In the same function, I noticed you use a try-catch block for underflow(). That ensures that it won't work with the given tester files, but that also ensures it'll work for other purposes. I'm guessing that's why you implemented it, but it does deviate from the ece 250 assignment. Just something to keep in mind. Not sure what our prof wants.